Renames modules to apis
