+++ server/routes/trips.js
@@ -1,7 +1,7 @@
 var express = require('express');
 var OpenRouter = express.Router();
 var AuthRouter = express.Router();
-var Trips = require('./../modules/trips');
+var Trips = require('./../apis/trips');
 
 AuthRouter.post('/addTrip', function (req, res) {
     Trips.addTrip(req.jwt.id, req.body, function (result) {
